+++ changes.md
@@ -3,6 +3,7 @@
 ## Changes in 5.1.x
 1. Fix - support newest TFormula in TF1 (#127)
 2. Fix - ignore NaN value in saved TF1 buffer
+3. Fix - correctly treat transparancy in geo painter
 
 
 ## Changes in 5.1.1