+++ changes.md
@@ -4,6 +4,7 @@
 1. Fix - support newest TFormula in TF1 (#127)
 2. Fix - ignore NaN value in saved TF1 buffer
 3. Fix - correctly treat transparancy in geo painter
+4. Fix - disable useFontCache for SVG mathjax output
 
 
 ## Changes in 5.1.1